Add files via upload
SVM Flower Image Classification
This project uses a Support Vector Machine (SVM) model to classify images of flowers into categories: Jasmine, Lily, and Sunflower. The images are preprocessed by resizing and flattening before being used to train the model. The performance is evaluated using accuracy, classification report, and confusion matrix. Additionally, it includes visualizations for model accuracy and confusion matrix, and it predicts the flower category for a random image.